requirements.txt
numpy
scipy
astropy
skycalc_cli
python-fsps
